feat: warmup in once
