+++ backend/groth16/bn254/icicle/device.go
@@ -1,0 +1,19 @@
+package icicle
+
+import (
+	"sync"
+
+	icicle_runtime "github.com/ingonyama-zk/icicle/v3/wrappers/golang/runtime"
+)
+
+var onceWarmUpDevice sync.Once
+
+func WarmUpDevice() {
+	onceWarmUpDevice.Do(func() {
+		icicle_runtime.LoadBackendFromEnvOrDefault()
+		device := icicle_runtime.CreateDevice("CUDA", 0)
+		icicle_runtime.RunOnDevice(&device, func(args ...any) {
+			icicle_runtime.WarmUpDevice()
+		})
+	})
+}
