+++ backend/groth16/bn254/icicle/provingkey.go
@@ -5,7 +5,6 @@ import (
 	groth16_bn254 "github.com/consensys/gnark/backend/groth16/bn254"
 	cs "github.com/consensys/gnark/constraint/bn254"
 	icicle_core "github.com/ingonyama-zk/icicle/v3/wrappers/golang/core"
-	icicle_runtime "github.com/ingonyama-zk/icicle/v3/wrappers/golang/runtime"
 )
 
 type deviceInfo struct {
@@ -22,14 +21,6 @@ type deviceInfo struct {
 type ProvingKey struct {
 	groth16_bn254.ProvingKey
 	*deviceInfo
-}
-
-func WarmUpDevice() {
-	icicle_runtime.LoadBackendFromEnvOrDefault()
-	device := icicle_runtime.CreateDevice("CUDA", 0)
-	icicle_runtime.RunOnDevice(&device, func(args ...any) {
-		icicle_runtime.WarmUpDevice()
-	})
 }
 
 func Setup(r1cs *cs.R1CS, pk *ProvingKey, vk *groth16_bn254.VerifyingKey) error {
